webasm: Add Engine interface to loop in emscripten
Emscripten main loops require changing the main thread of execution. A side
effect of this is that the calling thread does not resume its execution.
Conditional execution has been added to take care of this case in
emscripten while keeping behaviour the same with the use of engine
interface.

Lua Scripts now have 4 main functions with regards to execution:
* Initialize is run at scene load
* Terminate is run at scene unload
* Loop is run every frame

Things to consider in the future:
* Using multiple scripts will require load and unload to be moved to
  different calls.